* [sources]: fix test2 "JSON" timeouts during device setup
A profile can change after device setup. Thus the call to get the device
profile, which invokes command "setup" might not see the Oyranos specified
profile. To avoid the case of a device profile not reappearing in time,
the profile is passed internally from setup to the getter.

The bug was seen under a color server and possibly others moving the
profiles around. The wrong profile was returned during initial setup.