* [oall]: load only inbuild modules for static library
The oall code provides the glue to access the inbuild registered modules.
It initialises the oyCMMsGetLibNames_p and oyCMMinfoFromLibName_p
pointers with its own module registering.

The downside is, that outside build modules can not be loaded. They need
to be brought inside the code tree and linked in and registerd. All
dependencies are loaded at once.
On the upside, the library can now be used statical with hopefully fewer
load and install problems.