* [core]: add internal oyObjectStructTreeParentContains()
It tries to carefully avoid recursive calls. However endless recusive calls
might by possible. The new measures shall hint problems in this debug code
paths.

oyObjectStructTreeContains() uses contains_cache[] to show the call history
of the previous ID's in case of an error.

contains_nested variable together with the CRETURN macro try to detect
excessive recursive calls in oyObjectStructTreeContains(). In case of
more than 100 recursive calls it is used to print an error message.

oyObjectStructTreeParentContains() is used as well to detect upward circular
object connections.

In the center of oyObjectStructTreeContains() is an alternative approach to
avoid endless recursion. It uses the object ID history in contains_cache.