* [oydi]: disable local white point with no XCM
Disable linear effects with no XCM

Be carefule with both options to honour the edit flag and set
them modified automatic, for override on demand.

The case of correcting on the command line is not handled.
One option might be to add display_mode=1 to all viewers in
Fl_Oy_XXX and do not change options if that display_mode=1 option
is missed.

"oicc" is not a good place to handle display stuff. So all this is now
in "oydi".
There remains "white_point_adjust" effect profile creation in oicc,
which should move to oydi as well.

improves on #36