* [lcm2]: CGATS parsing is inconvinient with lcms API
there is no BEGIN_DATA_FORMAT/END_DATA_FORMAT API

tables are seen as static things in one file, which is in contrast of
in the field CGATS format use

Options are to write a parser from ground with libOyjl or
turn C++ CgatsFilter from ICCExamin into a Oyranos filter.
The later code is already shared, well tested and has all bells and whistles.