* [oyjl]: display command PNG output in oyjl-json-qml
The Process callback can obtain a PNG data stream from
e.g. oy-profile-graph -i A. That is encoded as base64 and
prefixed with a data URI. See:
https://en.wikipedia.org/wiki/Data_URI_scheme
The uri is later detected in the string and routed to the image
QML object. Otherwise the text is displayed as is in the helpTextArea.