Drop build host kernel version
This caused variations in build results
when building in a chroot or container on hosts with different
Linux kernel versions.

For openSUSE Tumbleweed we also update the kernel often,
but do not want to publish new oyranos binaries that
did not differ otherwise.

See also https://reproducible-builds.org/ for why this matters.