* [oyjl]: improve oyjlUi_ToMarkdown() options
make markdown OPTIONS output more of a table form to line down
the descriptions of options

Connect the option and argument by equal sign '=' in the SYNOPSIS section
for easier reading.

The equal sign is omitted in the OPTIONS section for making the belonging
to both, short and long options version understood and for line wrap.

As we have few control over CSS, the table layout is coded into the
elements for a uniform look.

The double minus'-' sign is preserved, both on the SYNOPSIS lines and
inside the HTML tables.