* [core]: use OpenIcc DB display/night on desktop
The goal is to display Material:System on desktop depending on
oyranos-monitor-white-point night setting. The dbus signal is
emitted by libelektra - hence the unfortune name space for detection.
libOpenIcc is used to parse the user OpenICC DB JSON file for the
night setting and checks if a night_effect is active. Currently there
is no opt-out of the night_effect possible, so the app uses on night
the night_effect to not interfere with its manipulations.