* [templates]: add oyStruct_RegisterStaticFreeFunc_f
... in order to release all static memory from object types. The
type specific init variable is reset by the object type itself.

oyStruct_RegisterStaticMessageFunc() API is changed and extented by
oyStruct_RegisterStaticFreeFunc_f argument, which replaces the
init variable adress. oyStruct_RegisterStaticMessageFunc() and
oyLibCoreRelease() are changed to reflect that API change.

Make all local static variables in type functions global. That is
used for release by the type specific oyStruct_RegisterStaticFreeFunc_f
functions in oyXXX_StaticFree_(). All customStaticMessage template blocks
are updated to reflect that change too.

Each class contains a oyXXX_StaticFree_() type specific deinitaliser,
which is passed into oyStruct_RegisterStaticMessageFunc() and called
inside oyLibCoreRelease().

Object allocators contain a switch in order to give valgrind a hint about
object ID's for debugging in the backtrace. It is only useable for low
object counts, but helps in nailing down test cases during development.

Beautify the debug messages in oyObject_Ref().

Fix observer related release issues and circular references:

Change the order of object ID and object->oy_ release.

Observer refs are counted twice for each connection, one for the
owned graph edge and one for the other objects graph edge.

In case of negative references, the object is considered in release state
and memory is not touched. The assumtion is, the object is touched by a
far relative parent/child relation and will be released in an other place
inside the function call stack.

Release references according to the observer count for the oyObject_s to
release it completly. The oyObject_s is disconnected from it's parent
object to avoid confusion durching child release.