* [core]: support custom objects in object tree API's
Affected public API is oyObjectTreePrint(), which shows the objec relation
in the DAG SVG graph.

oyStruct_GetId() is updated too.

Show implementation example in lcm2 module, with it's l2cmsProfileWrap_s type.

oyStruct_GetId() obtained a fix for empty graphs. It does omit writing to file
in such a case and gives a warning instead.