* [tools]: support HLC Atlas SVG/PNG from tree/NCC
The -i|--illuminant option is renamed to -u|--illuminant.

The tree can be generated by the --hlc options or
be imported with the new -i|--import option.

The flags symbol is renamed to not conflict with OYJL_IS_XXX macros.

A "Color Page" group is added.

Do not try oyImage_FromFile() with JSON data.

Put more page infos into the NCC output, in order to render the page
without any help from the --hlc variables.

pages data layout is slidly changed to add attributes. The colors index
is used to reference pathes instead of id. The layout orientation is top
left to buttom right. The JSON pages text give already a impression of
the rendered page.

The code to directly draw the --hlc page from its options is still in place.
But is no longer used and can be removed. So there is only one code path
for rendering both --hlc=NUMBER and -i=ncc_color_page.ncc options to SVG/PNG.