* [oyjl]: improve docu
The mandatory group field is a entry to mark and parse logical relations
of options and the group. But there is no powerful while simple enough
formal language for that yet, in order to make that part of parsing,
evaluation and so on. That field is solely used to emphasise these options.